Refator contents client test util
To reuse in new tests
