+++ jupyterfs/tests/test_fsmanager.py
@@ -6,18 +6,17 @@
 # the Apache License 2.0.  The full license can be found in the LICENSE file.
 
 from contextlib import nullcontext
-import json
-from pathlib import Path, PurePosixPath
+from pathlib import Path
 import pytest
 import os
 import shutil
 import socket
 
-from jupyter_core.utils import ensure_async
 import tornado.web
 
 from jupyterfs.fsmanager import FSManager
 from .utils import s3, samba
+from .utils.client import ContentsClient
 
 test_dir = "test"
 test_content = "foo\nbar\nbaz"
@@ -71,46 +70,6 @@ configs = [
         }
     },
 ]
-
-class ContentsClient:
-    def __init__(self, jp_fetch):
-        self.fetch = jp_fetch
-
-    async def set_resources(self, resources):
-        rep = await self.fetch(
-            "/jupyterfs/resources",
-            method='POST',
-            body=json.dumps({
-                'options': {},
-                'resources': resources,
-            })
-        )
-        return json.loads(rep.body)
-
-    async def mkdir(self, path, parents=False):
-        if parents:
-            pp = PurePosixPath(path)
-            for i, p in enumerate(pp.parts):
-                await self.mkdir("/".join(pp.parts[:i]))
-            return
-        rep = await self.fetch(
-            f"/api/contents/{path.strip('/')}",
-            method='PUT',
-            body=json.dumps({'type': 'directory'}),
-        )
-        return json.loads(rep.body)
-
-    async def save(self, path, model):
-        rep = await self.fetch(
-            f"/api/contents/{path}",
-            method='PUT',
-            body=json.dumps(model),
-        )
-        return json.loads(rep.body)
-
-    async def get(self, path):
-        rep = await self.fetch(f"/api/contents/{path}", raise_error=True)
-        return json.loads(rep.body)
 
 
 class _TestBase:
